moved fs functions to Filesystem
